Refactored routes to use express router for middleware
