+++ backend/services/passport.js
@@ -20,7 +20,7 @@ passport.use(
     {
       clientID: keys.googleClientID,
       clientSecret: keys.googleClientSecret,
-      callbackURL: '/auth/google/callback',
+      callbackURL: '/api/auth/google/callback',
       proxy: true
     },
     async (accessToken, refreshToken, profile, done) => {
