+++ client/package.json
@@ -28,9 +28,6 @@
     "eject": "react-scripts eject"
   },
   "proxy": {
-    "/auth": {
-      "target": "http://localhost:5000"
-    },
     "/api": {
       "target": "http://localhost:5000"
     }
