--- routes-new/api/event.js
@@ -1,19 +1,0 @@
-const router = require("express").Router();
-const eventController = require("../../controllers/eventController");
-
-router
-  .route("/")
-  .get(eventController.EXAMPLE)
-  .post(eventController.create);
-
-//.get(eventController.findAll)
-
-router
-  .route("/:id")
-  .get(eventController.findById)
-  .put(eventController.update)
-  .delete(eventController.remove);
-
-router.route("/search").post(eventController.search);
-
-module.exports = router;
