--- routes-new/api/search.js
@@ -1,7 +1,0 @@
-const router = require('express').Router();
-const searchController = require('../../controllers/searchController');
-
-router.route('/')
-  .post(searchController.search);
-
-module.exports = router;
