--- routes-new/api/track.js
@@ -1,19 +1,0 @@
-const router = require('express').Router();
-const trackController = require('../../controllers/trackController');
-
-router.route('/')
-  .get(trackController.findAll)
-  .post(trackController.create);
-
-router.route('/:id')
-  .get(trackController.findById)
-  .put(trackController.update)
-  .delete(trackController.remove);
-
-router.route('/search')
-  .post(trackController.search);
-
-router.route('/many')
-  .post(trackController.createMany);
-  
-module.exports = router;
