--- routes-new/api/venue.js
@@ -1,13 +1,0 @@
-const router = require('express').Router();
-const venueController = require('../../controllers/venueController');
-
-router.route('/')
-  .get(venueController.findAll)
-  .post(venueController.create);
-
-router.route('/:id')
-  .get(venueController.findById)
-  .put(venueController.update)
-  .delete(venueController.remove);
-
-module.exports = router;
